# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Technical Test Backend Golang",
        "title": "CAR RENTAL",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/driver-incentives": {
            "get": {
                "description": "Retrieve a list of all driver incentives.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "driverIncentives"
                ],
                "summary": "Retrieve list of driver incentives",
                "responses": {
                    "200": {
                        "description": "List of driver incentives",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.DriverIncentive"
                            }
                        }
                    },
                    "404": {
                        "description": "No driver incentive found",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Add a new driver incentive record.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "driverIncentives"
                ],
                "summary": "Create a new driver incentive",
                "parameters": [
                    {
                        "description": "Driver Incentive Data",
                        "name": "driverIncentive",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.InputDriverIncentive"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created driver incentive",
                        "schema": {
                            "$ref": "#/definitions/models.DriverIncentive"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/driver-incentives/driver/{id}": {
            "get": {
                "description": "Retrieve a list of incentives for a given driver ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Drivers Incentive"
                ],
                "summary": "Get Driver Incentives List",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Driver ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of Incentives",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.DriverIncentive"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid Driver ID",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No driverIncentive found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Driver not found / Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/driver-incentives/driver/{id}/total": {
            "get": {
                "description": "Retrieve the total incentive amount for a given driver ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Drivers Incentive"
                ],
                "summary": "Get Total Driver Incentives",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Driver ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Total Incentive for Driver",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid Driver ID",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Driver not found / Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/driver-incentives/{id}": {
            "get": {
                "description": "Retrieve a specific driver incentive by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "driverIncentives"
                ],
                "summary": "Retrieve a driver incentive by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Driver Incentive ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Driver incentive details",
                        "schema": {
                            "$ref": "#/definitions/models.DriverIncentive"
                        }
                    },
                    "400": {
                        "description": "Invalid driver incentive ID",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Driver incentive not found",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Modify details of an existing driver incentive.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "driverIncentives"
                ],
                "summary": "Update a driver incentive",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Driver Incentive ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated driver incentive data",
                        "name": "driverIncentive",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.InputDriverIncentive"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated driver incentive",
                        "schema": {
                            "$ref": "#/definitions/models.DriverIncentive"
                        }
                    },
                    "400": {
                        "description": "Invalid required param or request body",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Driver incentive not found",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove a driver incentive using its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "driverIncentives"
                ],
                "summary": "Delete a driver incentive by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Driver Incentive ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful deletion message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid required param",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Driver incentive not found",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/pkg.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Booking": {
            "type": "object",
            "properties": {
                "book_type_id": {
                    "type": "integer"
                },
                "booking_type": {
                    "$ref": "#/definitions/models.BookingType"
                },
                "car": {
                    "$ref": "#/definitions/models.Car"
                },
                "car_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "customer": {
                    "$ref": "#/definitions/models.Customer"
                },
                "customer_id": {
                    "type": "integer"
                },
                "discount": {
                    "type": "integer"
                },
                "driver": {
                    "$ref": "#/definitions/models.Driver"
                },
                "driver_id": {
                    "type": "integer"
                },
                "end_rent": {
                    "type": "string"
                },
                "finished": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "start_rent": {
                    "type": "string"
                },
                "total_cost": {
                    "type": "integer"
                },
                "total_driver_cost": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.BookingType": {
            "type": "object",
            "properties": {
                "booking_type": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Car": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "daily_rent": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "stock": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Customer": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "membership": {
                    "$ref": "#/definitions/models.Membership"
                },
                "membership_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "nik": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Driver": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "daily_cost": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "nik": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.DriverIncentive": {
            "type": "object",
            "properties": {
                "booking": {
                    "$ref": "#/definitions/models.Booking"
                },
                "booking_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "incentive": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.InputDriverIncentive": {
            "type": "object",
            "required": [
                "booking_id",
                "incentive"
            ],
            "properties": {
                "booking_id": {
                    "type": "integer"
                },
                "incentive": {
                    "type": "integer"
                }
            }
        },
...
        "models.Membership": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "discount": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "membership_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "pkg.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}